Able to parse most cube information
Can parse vertex and cube data, but is not able to handle levels with walls, due to a bug in the wall parsing. When walls are present, it causes the parsing of later cubes to be incorrect. So, parsing hall.rdl works fine, but parsing shortHall.rdl does not work.

Includes parsing of primary and secondary textures